+++ src/main/java/ch/ethz/matsim/discrete_mode_choice/modules/ModelModule.java
@@ -43,6 +43,7 @@ public class ModelModule extends AbstractModule {
 		install(new TourFinderModule());
 		install(new SelectorModule());
 		install(new ConstraintModule());
+		install(new FilterModule());
 
 		bind(ModeChainGeneratorFactory.class).to(DefaultModeChainGenerator.Factory.class);
 	}