+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-**1.0.9-dev**
+**1.0.9**
 
 - Make PlanBuilder consistent with SubtourModeChoice
 - Improve ModeAvailability documentation